Déplace la route "/utilisateur/edition" vers le routeur dédié
