+++ src/routes/connecte/routesConnectePage.js
@@ -1,26 +1,48 @@
 const express = require('express');
 
-const routesConnectePage = ({ depotDonnees }) => {
+const routesConnectePage = ({ middleware, depotDonnees, referentiel }) => {
   const routes = express.Router();
 
-  routes.get('/motDePasse/edition', (requete, reponse) => {
-    const idUtilisateur = requete.idUtilisateurCourant;
-    depotDonnees.utilisateur(idUtilisateur).then((utilisateur) =>
-      reponse.render('motDePasse/edition', {
-        utilisateur,
-        afficheChallengeMotDePasse: true,
-      })
-    );
-  });
-
-  routes.get('/motDePasse/initialisation', (requete, reponse) => {
-    const idUtilisateur = requete.idUtilisateurCourant;
-    depotDonnees
-      .utilisateur(idUtilisateur)
-      .then((utilisateur) =>
-        reponse.render('motDePasse/edition', { utilisateur })
+  routes.get(
+    '/motDePasse/edition',
+    middleware.verificationJWT,
+    (requete, reponse) => {
+      const idUtilisateur = requete.idUtilisateurCourant;
+      depotDonnees.utilisateur(idUtilisateur).then((utilisateur) =>
+        reponse.render('motDePasse/edition', {
+          utilisateur,
+          afficheChallengeMotDePasse: true,
+        })
       );
-  });
+    }
+  );
+
+  routes.get(
+    '/motDePasse/initialisation',
+    middleware.verificationJWT,
+    (requete, reponse) => {
+      const idUtilisateur = requete.idUtilisateurCourant;
+      depotDonnees
+        .utilisateur(idUtilisateur)
+        .then((utilisateur) =>
+          reponse.render('motDePasse/edition', { utilisateur })
+        );
+    }
+  );
+
+  routes.get(
+    '/utilisateur/edition',
+    middleware.verificationJWT,
+    (requete, reponse) => {
+      const departements = referentiel.departements();
+      const idUtilisateur = requete.idUtilisateurCourant;
+      depotDonnees
+        .utilisateur(idUtilisateur)
+        .then((utilisateur) =>
+          reponse.render('utilisateur/edition', { utilisateur, departements })
+        );
+    }
+  );
 
   return routes;
 };
